git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/runtime_verification@15609 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/cpa/smg/SMG.java
@@ -874,7 +874,7 @@ final class NeqRelation {
 
   public void add_relation(Integer pOne, Integer pTwo) {
 
-    if(pOne == pTwo) {
+    if(pOne.intValue() == pTwo.intValue()) {
       return;
     }
